improve documentation
A few errors were made explaining the `skew()` and `translate()`
shorthand methods, due to originally copy pasting the `scale()`
shorthand method.

The first argument of `scale()` applies to both X and Y-axis,
where the optional second argument overrides the Y-axis.

The first argument of `skew()` and `translate()` however only
apply to the X-axis, and the optional second argument defines
the Y-axis.

This is how CSS works, and how Rematrix works... but the docs
incorrectly described the same behavior as `scale()` for
`skew()` and `translate()`.